database/requirements.txt
django-easy-pdf==0.1.0
django-xhtml2pdf==0.0.3
reportlab==3.3.0
